+++ container-engine-lib/lib/backend_impls/metrics_reporting/metrics_reporting_kurtosis_backend.go
@@ -7,6 +7,8 @@ import (
 	"github.com/kurtosis-tech/container-engine-lib/lib/backend_interface/objects/enclave"
 	"github.com/kurtosis-tech/container-engine-lib/lib/backend_interface/objects/engine"
 	"github.com/kurtosis-tech/container-engine-lib/lib/backend_interface/objects/exec_result"
+	"github.com/kurtosis-tech/container-engine-lib/lib/backend_interface/objects/file_artifact"
+	"github.com/kurtosis-tech/container-engine-lib/lib/backend_interface/objects/file_artifact_expansion_volume"
 	"github.com/kurtosis-tech/container-engine-lib/lib/backend_interface/objects/module"
 	"github.com/kurtosis-tech/container-engine-lib/lib/backend_interface/objects/networking_sidecar"
 	"github.com/kurtosis-tech/container-engine-lib/lib/backend_interface/objects/port_spec"
@@ -553,4 +555,37 @@ func (backend *MetricsReportingKurtosisBackend) DestroyNetworkingSidecars(
 		return nil, nil, stacktrace.Propagate(err, "An error occurred destroying networking sidecars using filters '%+v'", filters)
 	}
 	return successfulUserServiceGuids, erroredUserServiceGuids, nil
+}
+
+func (backend *MetricsReportingKurtosisBackend) CreateFileArtifactExpansionVolume(
+	ctx context.Context,
+	enclaveId enclave.EnclaveID,
+	serviceGuid service.ServiceGUID,
+	fileArtifactId file_artifact.FilterArtifactID,
+)(
+	*file_artifact_expansion_volume.FileArtifactExpansionVolume,
+	error,
+) {
+	newFileArtifactExpansionVolume, err := backend.underlying.CreateFileArtifactExpansionVolume(ctx, enclaveId, serviceGuid, fileArtifactId)
+	if err != nil {
+		return nil, stacktrace.Propagate(err, "An error occurred creating file artifact expansion volume for user service with GUID '%v' and file artifact ID '%v' in enclave with ID '%v'", serviceGuid, fileArtifactId, enclaveId)
+	}
+
+	return newFileArtifactExpansionVolume, nil
+}
+
+func (backend *MetricsReportingKurtosisBackend) DestroyFileArtifactExpansionVolumes(
+	ctx context.Context,
+	filters *file_artifact_expansion_volume.FileArtifactExpansionVolumeFilters,
+) (
+	map[file_artifact_expansion_volume.FileArtifactExpansionVolumeName]bool,
+	map[file_artifact_expansion_volume.FileArtifactExpansionVolumeName]error,
+	error,
+) {
+	successfulExpansionVolumeNames, erroredExpansionVolumeNames, err := backend.underlying.DestroyFileArtifactExpansionVolumes(ctx, filters)
+	if err != nil {
+		return nil, nil, stacktrace.Propagate(err, "An error occurred destroying file artifact expansion volumes using filters '%+v'", filters)
+	}
+
+	return successfulExpansionVolumeNames, erroredExpansionVolumeNames, nil
 }